+++ lib/routes/pages/entry/new_entry_page.dart
@@ -239,8 +239,8 @@ class EntryValueTextField extends StatelessWidget {
       //TODO add some sort of message for user to fill in value and/or category
       return;
     }
-    int entryValue =
-        int.parse(value) * (categoryNotifier.value!.isExpense ? -1 : 1);
+    double entryValue =
+        double.parse(value) * (categoryNotifier.value!.isExpense ? -1 : 1);
     Entry entry = Entry(value: entryValue, dateTime: DateTime.now());
     EntryController.addEntry(entry, categoryNotifier.value!);
     Navigator.pop(context);
